+++ packages/button/stories/ButtonGroup.stories.tsx
@@ -50,6 +50,7 @@ const buttonTemplateWithStates: Decorator = (storyComponent, context) => {
 export default {
 	component: ButtonGroup,
 	title: 'Legacy/Button/ButtonGroup',
+	tags: ['deprecated'],
 	description: 'ButtonGroups group related actions and trigger them based on user interaction.',
 	decorators: [buttonTemplateWithStates],
 	parameters: {