+++ packages/components/stories/LinkIconButton.stories.tsx
@@ -5,6 +5,12 @@ import { LinkIconButton } from '../src/LinkIconButton';
 const meta: Meta<typeof LinkIconButton> = {
 	component: LinkIconButton,
 	title: 'Components/Navigation/LinkIconButton',
+	parameters: {
+		figma: {
+			design:
+				'https://www.figma.com/design/98HKKXL2dTle29ikJ3tzk7/%F0%9F%9A%80-LaunchPad?node-id=8079-92746&m=dev',
+		},
+	},
 };
 
 export default meta;