+++ packages/filter/stories/FilterButton.stories.tsx
@@ -5,6 +5,7 @@ import { FilterButton } from '../src';
 export default {
 	component: FilterButton,
 	title: 'Legacy/Filter/FilterButton',
+	tags: ['deprecated'],
 	description:
 		"When the Filter component doesn't suffice, construct your own custom filter using the FilterButton.",
 	parameters: {